+++ cmd/vegawallet/commands/service_run.go
@@ -459,13 +459,11 @@ func handleAPIv2Request(interaction interactor.Interaction, responseChan chan<- 
 		}
 		p.Print(str)
 		selectedWallet := readInput(p.String().QuestionMark().Text("Which wallet do you want to use? "), p, data.AvailableWallets)
-		passphrase := readPassphrase(p.String().BlueArrow().Text("Enter the passphrase for the wallet \"").InfoText(selectedWallet).Text("\": "), p)
 		responseChan <- interactor.Interaction{
 			TraceID: interaction.TraceID,
 			Name:    interactor.SelectedWalletName,
 			Data: interactor.SelectedWallet{
-				Wallet:     selectedWallet,
-				Passphrase: passphrase,
+				Wallet: selectedWallet,
 			},
 		}
 	case interactor.RequestPassphrase: